# Conflicts:
#	frontend/src/app/modules/turma/components/turma-form/turma-form.component.ts

+++ backend/src/main/java/com/basis/sgcproject/service/TurmaFormacaoService.java
@@ -71,6 +71,10 @@ public class TurmaFormacaoService {
 
     @Transactional
     public void excluir(Integer turmaFormacaoId) {
+        TurmaFormacao turma = turmaFormacaoRepository.getById(turmaFormacaoId);
+        if (turma.getStatus().getDescricao().equals("Iniciada")) {
+            throw new RegraNegocioException("Esta turma não pode ser excluida pois está em andamento.");
+        }
         turmaFormacaoRepository.deleteById(turmaFormacaoId);
     }
 }
